+++ src/styled/Label/tests/Label.test.js
@@ -1,11 +1,16 @@
-// import React from 'react';
-// import { mount } from 'enzyme';
-// import { enzymeFind } from 'styled-components/test-utils';
+import React from 'react';
+import { shallow } from 'enzyme';
 
-// import Label from '../index';
+import Label from '../index';
+
+const renderComponent = () => {
+  const wrapper = shallow(<Label />);
+
+  return wrapper;
+};
 
 describe('<Label />', () => {
-  it('Expect to have unit tests specified', () => {
-    expect(true).toEqual(true);
+  it('should not crash', () => {
+    renderComponent();
   });
 });
