+++ stories/Compos/Select.stories.js
@@ -1,18 +1,17 @@
 import React from 'react';
 /* eslint-disable import/no-extraneous-dependencies */
 import { storiesOf } from '@storybook/react';
-import Textarea from '../../src/components/Textarea';
+import Select from '../../src/components/Select';
 
 const defaultProps = {
-  placeholder: 'Write your little description here…',
-  name: 'textarea',
-  rows: '4',
+  name: 'select',
+  options: ['first option', 'second option'],
 };
 
-class TextareaStory extends React.Component {
+class SelectStory extends React.Component {
   state = {
     value: {
-      textarea: '',
+      select: '',
     },
   };
 
@@ -26,13 +25,13 @@ class TextareaStory extends React.Component {
     const { value } = this.state;
 
     return (
-      <Textarea
+      <Select
         {...defaultProps}
         onChange={this.handleChange}
-        value={value.textarea}
+        value={value.select}
       />
     );
   }
 }
 
-storiesOf('Components|Textarea', module).add('Simple', () => <TextareaStory />);
+storiesOf('Components|Select', module).add('Simple', () => <SelectStory />);
