+++ stories/Styled/Select.stories.js
@@ -1,0 +1,6 @@
+import React from 'react';
+/* eslint-disable import/no-extraneous-dependencies */
+import { storiesOf } from '@storybook/react';
+import Select from '../../src/styled/Select';
+
+storiesOf('Styled|Select', module).add('Simple', () => <Select />);
